2392: extract comment step into its own compoent
